Moved AutoFileNames out of AtomDownloadJob and added a test.
